requirements.txt
gitpython == 3.1.27
bottle == 0.12.25
Faker == 24.14.1
icecream == 2.1.3
requests
